Massive commit. Rework of entire directory structure
Moved the junk files we had at the root level to a folder labeled 'unsure', nested in the history folder. Was a bit of an eyesore. Also, the entire Django project is now located within the academic-pathway-dev folder. Please be careful here; we have swapped to SQLite, which means that the database is checked into version control. If you pull and make changes to the DB locally, then push, they will be reflected in the live environment.